+++ plugins/gh-datainmap/CHANGELOG.md
@@ -1,3 +1,5 @@
+- Filters omgezet van een `aside`-element naar een `section`-element.
+
 **1.8.2 (2020-09-23)**
 
 - Kaart krijgt nu automatisch focus bij een mouseenter-event. Hierdoor hoeft bij gebruik van de muis niet eerst op de kaart geklikt te worden. Bij een mouseleave-event vervalt de focus weer.